delete obsolete files and rename vectorized implementations
