+++ lukefi/metsi/domain/natural_processes/grow_acta.py
@@ -1,8 +1,6 @@
-from lukefi.metsi.app.utils import MetsiException
-from lukefi.metsi.forestry.naturalprocess.grow_acta import grow_diameter_and_height, grow_diameter_and_height_vectorized
 from lukefi.metsi.data.model import ForestStand, ReferenceTree
-
-from lukefi.metsi.domain.natural_processes.util import update_stand_growth, update_stand_growth_vectorized
+from lukefi.metsi.domain.natural_processes.util import update_stand_growth
+from lukefi.metsi.forestry.naturalprocess.grow_acta import grow_diameter_and_height
 from lukefi.metsi.sim.collected_data import OpTuple
 
 
@@ -18,24 +16,10 @@ def split_sapling_trees(trees: list[ReferenceTree]) -> tuple[list[ReferenceTree]
 
 def grow_acta(input_: OpTuple[ForestStand], /, **operation_parameters) -> OpTuple[ForestStand]:
     step = operation_parameters.get('step', 5)
-    stand, _ = input_
-    if len(stand.reference_trees) == 0:
+    stand, collected_data = input_
+    if stand.reference_trees.size == 0:
         return input_
     diameters, heights = grow_diameter_and_height(stand.reference_trees, step)
-    stems = list(map(lambda x: x.stems_per_ha, stand.reference_trees))
+    stems = stand.reference_trees.stems_per_ha
     update_stand_growth(stand, diameters, heights, stems, step)
-    return stand, input_[1]
-
-
-
-def grow_acta_vectorized(input_: OpTuple[ForestStand], /, **operation_parameters) -> OpTuple[ForestStand]:
-    step = operation_parameters.get('step', 5)
-    stand, collected_data = input_
-    if stand.reference_trees_soa is None:
-        raise MetsiException("Reference trees not vectorized")
-    if stand.reference_trees_soa.size == 0:
-        return input_
-    diameters, heights = grow_diameter_and_height_vectorized(stand.reference_trees_soa, step)
-    stems = stand.reference_trees_soa.stems_per_ha
-    update_stand_growth_vectorized(stand, diameters, heights, stems, step)
     return stand, collected_data
