+++ lukefi/metsi/forestry/naturalprocess/grow_acta.py
@@ -1,13 +1,12 @@
-from collections import defaultdict
 import math
-from statistics import median
 import numpy as np
 import numpy.typing as npt
 
-from lukefi.metsi.data.model import ReferenceTree, TreeSpecies
+from lukefi.metsi.data.model import TreeSpecies
 from lukefi.metsi.data.vector_model import ReferenceTrees
 
 
+@np.vectorize
 def yearly_diameter_growth_by_species(
     spe: TreeSpecies,
     d: float,
@@ -38,6 +37,7 @@ def yearly_diameter_growth_by_species(
     return growth_percent
 
 
+@np.vectorize
 def yearly_height_growth_by_species(
     spe: TreeSpecies,
     d: float,
@@ -64,50 +64,9 @@ def yearly_height_growth_by_species(
     return growth_percent
 
 
-def grow_diameter_and_height(
-    trees: list[ReferenceTree],
-    step: int = 5
-) -> tuple[list[float], list[float]]:
-    """ Diameter and height growth for trees with height > 1.3 meters. Based on Acta Forestalia Fennica 163. """
-    if not trees:
-        return [], []
-    group = defaultdict(list)
-    for i, t in enumerate(trees):
-        group[t.species].append(i)
-    ds = [t.breast_height_diameter or 0 for t in trees]
-    hs = [t.height for t in trees]
-    for s in range(step):
-        bigh = [h for h in hs if h >= 1.3]
-        if bigh:
-            hdom = median(bigh)
-            gs = [t.stems_per_ha * math.pi * (0.01 * 0.5 * d)**2 for t, d in zip(trees, ds)]
-            g = sum(gs)
-            for spe, idx in group.items():
-                gg = sum(gs[i] for i in idx)
-                ag = sum((trees[i].biological_age + s) * gs[i] for i in idx) / gg
-                dg = sum(ds[i] * gs[i] for i in idx) / gg
-                hg = sum(hs[i] * gs[i] for i in idx) / gg
-                for i in idx:
-                    if hs[i] >= 1.3:
-                        pd = yearly_diameter_growth_by_species(spe, ds[i], hs[i], ag, dg, hg, hdom, g) / 100
-                        ph = yearly_height_growth_by_species(spe, ds[i], hs[i], ag, dg, hg, g) / 100
-                        ds[i] *= 1 + pd
-                        hs[i] *= 1 + ph
-        for i, h in enumerate(hs):
-            if h < 1.3:
-                hs[i] += 0.3
-                if hs[i] >= 1.3 and not ds[i]:
-                    ds[i] = 1.0
-    return ds, hs
-
-
-yearly_diameter_growth_by_species_vectorized = np.vectorize(yearly_diameter_growth_by_species)
-yearly_height_growth_by_species_vectorized = np.vectorize(yearly_height_growth_by_species)
-
-
-def grow_diameter_and_height_vectorized(trees: ReferenceTrees,
-                                        step: int = 5) -> tuple[npt.NDArray[np.float64],
-                                                                npt.NDArray[np.float64]]:
+def grow_diameter_and_height(trees: ReferenceTrees,
+                             step: int = 5) -> tuple[npt.NDArray[np.float64],
+                                                     npt.NDArray[np.float64]]:
     """
     Diameter and height growth for trees with height > 1.3 meters. Based on Acta Forestalia Fennica 163.
     Vector data implementation.
@@ -133,21 +92,21 @@ def grow_diameter_and_height_vectorized(trees: ReferenceTrees,
 
                 condition = (trees.species == spe) & (hs >= 1.3)
 
-                pd = yearly_diameter_growth_by_species_vectorized(spe,
-                                                                  ds[condition],
-                                                                  hs[condition],
-                                                                  ag,
-                                                                  dg,
-                                                                  hg,
-                                                                  hdom,
-                                                                  g) / 100
-                ph = yearly_height_growth_by_species_vectorized(spe,
-                                                                ds[condition],
-                                                                hs[condition],
-                                                                ag,
-                                                                dg,
-                                                                hg,
-                                                                g) / 100
+                pd = yearly_diameter_growth_by_species(spe,
+                                                       ds[condition],
+                                                       hs[condition],
+                                                       ag,
+                                                       dg,
+                                                       hg,
+                                                       hdom,
+                                                       g) / 100
+                ph = yearly_height_growth_by_species(spe,
+                                                     ds[condition],
+                                                     hs[condition],
+                                                     ag,
+                                                     dg,
+                                                     hg,
+                                                     g) / 100
                 ds[condition] *= (1 + pd)
                 hs[condition] *= (1 + ph)
         hs[hs < 1.3] += 0.3
